+++ src/models/index.ts
@@ -1,2 +1,3 @@
-export * from './EventData';
-export * from './MessageHandlerData';
+export * from "./EventData";
+export * from "./MessageHandlerData";
+export * from "./Theme";